Seperated os dependencies from python/app dependencies from install.sh. OS dependencies now need to be install before running install.sh

+++ install.sh
@@ -7,37 +7,41 @@ APP_HOME="${INSTALL_ROOT}/metrilyx";
 if [[ -f "/etc/redhat-release" ]]; then
 	HTTPD="httpd"
 	HTTP_USER="apache"
-	PKGS="libuuid gcc uuid ${HTTPD} mod_wsgi python-setuptools python-devel mongo-10gen mongo-10gen-server"
-	PKG_INSTALLER="yum -y install"
-	PKG_LISTER="rpm -qa"
-	PKG_S_PREFIX="^"
+	#PKGS="libuuid gcc uuid ${HTTPD} mod_wsgi python-setuptools python-devel mongo-10gen mongo-10gen-server"
+	#PKG_INSTALLER="yum -y install"
+	#PKG_LISTER="rpm -qa"
+	#PKG_S_PREFIX="^"
 elif [[ -f "/etc/debian_version" ]]; then
 	HTTPD="apache2"
 	HTTP_USER="www-data"
-	PKGS="libuuid1 gcc uuid ${HTTPD} libapache2-mod-wsgi python-setuptools python-dev mongodb mongodb-server"
-	PKG_INSTALLER="apt-get install -y"
-	PKG_LISTER="dpkg -l"
-	PKG_S_PREFIX="ii\s+"
+	#PKGS="libuuid1 gcc uuid ${HTTPD} libapache2-mod-wsgi python-setuptools python-dev mongodb mongodb-server"
+	#PKG_INSTALLER="apt-get install -y"
+	#PKG_LISTER="dpkg -l"
+	#PKG_S_PREFIX="ii\s+"
 else
 	echo "Currently only RedHat/Debian based distro are supported.  Please install manually.";
 	exit 1;
 fi
 
 clean() {
-	find . -name '*.pyc' -exec rm -rvf '{}' \;
+	find . -name '*.pyc' -exec rm -rf '{}' \;
 }
-setup_app_dirs() {
+
+install_app() {
 	mkdir -p ${APP_HOME};
 	cp -a . ${APP_HOME}/;
 	chmod g+w ${APP_HOME};
 	( id celery 2>&1 ) > /dev/null || useradd celery;
 	chgrp celery ${APP_HOME};
 }
+
 setup_celery_startup() {
-	cp -a etc/rc.d/init.d/* /etc/rc.d/init.d/;
-	if [ ! -f /etc/sysconfig/celeryd ]; then 
-		cp etc/sysconfig/celeryd /etc/sysconfig/;
-	fi	
+	if [[ -f "/etc/redhat-release" ]]; then
+		cp -a etc/rc.d/init.d/* /etc/rc.d/init.d/;
+		if [ ! -f /etc/sysconfig/celeryd ]; then 
+			cp etc/sysconfig/celeryd /etc/sysconfig/;
+		fi	
+	fi
 }
 
 install_pydeps() {
@@ -50,43 +54,31 @@ install_pydeps() {
 
 backup_curr_install() {
 	clean;
-	if [ -d "${INSTALL_ROOT}/metrilyx" ]; then
+	if [ -d "${APP_HOME}" ]; then
 		echo "- Backing up existing installation...";
 		mv ${APP_HOME} ${APP_HOME}-${INSTALL_TIME};
 	fi;
 }
-setup_app_config() {
-	if [ -f "/opt/metrilyx-${INSTALL_TIME}/etc/metrilyx/metrilyx.conf" ]; then
-		echo "- Importing existing data..."
-		echo "  configs...";
+configure_app() {
+	echo "- Importing existing data..."
+	echo "  configs...";
+	if [ -f "${APP_HOME}-${INSTALL_TIME}/etc/metrilyx/metrilyx.conf" ]; then
 		cp ${APP_HOME}-${INSTALL_TIME}/etc/metrilyx/metrilyx.conf ${APP_HOME}/etc/metrilyx/metrilyx.conf;
-		if [ -f "${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js" ]; then
-			cp ${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js ${APP_HOME}/metrilyx/static/config.js;
-		fi
-		if [ ! -f "${APP_HOME}/metrilyx/static/config.js" ]; then
-			cp ${APP_HOME}/metrilyx/static/config.js.sample ${APP_HOME}/metrilyx/static/config.js;
-		fi
-		echo "  dashboards..."
-		cp -a ${APP_HOME}-${INSTALL_TIME}/pagemodels ${APP_HOME}/;
-		echo "  heatmaps..."
-		cp -a ${APP_HOME}-${INSTALL_TIME}/heatmaps ${APP_HOME}/;
 	else
-		cp ${APP_HOME}/metrilyx/static/config.js.sample ${APP_HOME}/metrilyx/static/config.js;
 		cp etc/metrilyx/metrilyx.conf.sample ${APP_HOME}/etc/metrilyx/metrilyx.conf;
 		${EDITOR:-vi} ${APP_HOME}/etc/metrilyx/metrilyx.conf;
+	if 
+	if [ -f "${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js" ]; then
+		cp ${APP_HOME}-${INSTALL_TIME}/metrilyx/static/config.js ${APP_HOME}/metrilyx/static/config.js;
+	else
+		cp ${APP_HOME}/metrilyx/static/config.js.sample ${APP_HOME}/metrilyx/static/config.js;
 	fi
-}
-install_app(){
-	echo "- Installing app..."
-	setup_app_dirs;
-	setup_app_config;
+
+	echo "  dashboards..."
+	cp -a ${APP_HOME}-${INSTALL_TIME}/pagemodels ${APP_HOME}/;
 	
-}
-app_postinstall() {
-	if [[ -f "/etc/redhat-release" ]]; then
-		setup_celery_startup;
-	fi
-	# apache restart
+	echo "  heatmaps..."
+	cp -a ${APP_HOME}-${INSTALL_TIME}/heatmaps ${APP_HOME}/;
 }
 configure_apache() {
 	echo "- Installing web components..."
@@ -121,21 +113,13 @@ if [ "$(whoami)" != "root" ]; then
 	exit 1;
 fi
 
-if [ "$1" == "all" ]; then
+if [ "$1" == "app" ]; then
 	install_pydeps;
 	backup_curr_install;
 	install_app;
+	configure_app;
 	configure_apache;
-	app_postinstall;
-	echo "(todo): Install and init postgres"
-	#init_postgres && init_django;
-	# apache restart
-elif [ "$1" == "app" ]; then
-	install_pydeps;
-	backup_curr_install;
-	install_app;
-	configure_apache;
-	app_postinstall;
+	setup_celery_startup;
 else
 	echo "Executing $1...";
 	$1;
@@ -143,7 +127,6 @@ fi
 
 
 echo ""
-echo " ** Heatmaps are still in beta phase, currently requiring a frequent restart."
 echo " ** If you choose to use heatmaps set the config options"
 echo " ** (/opt/metrilyx/etc/metrilyx/metrilyx.conf) and start celerybeat and celeryd."
 echo ""
